Bison Resting
The bison now has an energy value that drains every second by a set value. Once the energy reaches 0, the boolean "Is Tired" will set to true. This then calls an event that stops the movement of the bison and lets it regain energy whilst it is no longer moving. There is a random int in range that decides how long the bison will rest for, currently it is set from 10 to 20 seconds, this is only to test the feature and make sure it works