Bison damage and death
Bison can now be damaged by the weapons in game, there is set values for all weapons in the bison code but i can later change it to inherit from the weapons blueprint if we add more weapons.

When the animal has taken damage, it will begin to run away, being more realistic than just continuing its normal movement. It will do this until it has run far enough from the radius set and will then continue its normal movement again.

When the animal dies it will stop moving, i plan to add a death animation and then have the body be able to be looted, but for now it will despawn after 5-10 seconds