Inventory
Inventory now works as it should. Player can pick up items as long as they have a ItemData component attached. Inventory opens with "I" and you can drag and drop items across the inventory. There is also icons that show what the item is aswell as a quantity that shows below the icon.