Chests and looting bison
Chests can now be accessed and contain items that the player can take from and put into their own inventory. When a bison is killed, its corpse can be looted and it will have its own inventory system which works the same was as a chest. i also fixed an old bug with my rituals which made the interact key unusable after using the ritual site. (oli this might also fix the problem thats occuring on your branch)