Home for animals
the hunting animals now have a home at which they can spawn.
Also when an animal is tired, rather than resting in a random place it will walk back home to rest. This can later be expanded into the human AI where the people in our town can rest in the teepees